New Program, Filler classes, add and delete methods in channels.py
