+++ dizqueTV/helpers.py
@@ -4,7 +4,7 @@ from dizqueTV.exceptions import MissingSettingsError
 
 def combine_settings(new_settings_dict: json, old_settings_dict: json) -> json:
     """
-    Build a complete dictionary for new settings
+    Build a complete dictionary for new settings, using old settings as a base
     :param new_settings_dict: Dictionary of new settings kwargs
     :param old_settings_dict: Current settings
     :return: Dictionary of new settings
